+++ web/src/Layout/AppNav/VerticalNavWrapper.js
@@ -4,7 +4,7 @@ import {withRouter} from 'react-router-dom';
 import MetisMenu from 'react-metismenu';
 
 import {MainNav, ComponentsNav, FormsNav, WidgetsNav, ChartsNav, ExampleMainNav,
-ContactDoctorNav, PillsNav, ScheduleNav} from './NavItems';
+ContactDoctorNav, PillsNav, ScheduleNav, CheckUserAnalysisNav, SetUserScheduleNav, MessageUsersNav} from './NavItems';
 
 import * as firebase from "firebase/app";
 import "firebase/database";
@@ -13,15 +13,21 @@ class Nav extends Component {
     constructor(props) {
         super(props);
         this.state = {
-            isDeveloper: true
+            isDeveloper: true,
+            isCaregiver: true,
         }
     }
 
     componentDidMount() {
         if(firebase.auth().currentUser) {
-            this.setState({
-                isDeveloper: false,
-            })
+            let userId = firebase.auth().currentUser.uid;
+            firebase.database().ref('/users/'+userId).once('value')
+            .then(snapshot => {
+                this.setState({
+                    isDeveloper: false,
+                    isCaregiver: snapshot.val().isCaregiver,
+                })
+            });
         }
     }
 
@@ -31,12 +37,33 @@ class Nav extends Component {
         return (
             <Fragment>
                 <h5 className="app-sidebar__heading">Health Monitoring System</h5>
-                <MetisMenu content={MainNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
-                <MetisMenu content={PillsNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
-                <MetisMenu content={ScheduleNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
-                <MetisMenu content={ContactDoctorNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                {/* Cargiver View */}
+                {this.state.isCaregiver &&
+                <div>
+                    <MetisMenu content={CheckUserAnalysisNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={SetUserScheduleNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={MessageUsersNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                </div>
+                }
+                {/* Patient View */}
+                {!this.state.isCaregiver &&
+                <div>
+                    <MetisMenu content={MainNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={PillsNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={ScheduleNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={ContactDoctorNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                </div>
+                }
+                {/* Developer View */}
                 {this.state.isDeveloper &&
                 <div>
+                    <MetisMenu content={CheckUserAnalysisNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={SetUserScheduleNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={MessageUsersNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={MainNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={PillsNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={ScheduleNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
+                    <MetisMenu content={ContactDoctorNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
                     <h5 className="app-sidebar__heading">Example Menu</h5>
                     <MetisMenu content={ExampleMainNav} activeLinkFromLocation className="vertical-nav-menu" iconNamePrefix="" classNameStateIcon="pe-7s-angle-down"/>
                     <h5 className="app-sidebar__heading">Example UI Components</h5>