+++ web/Changelog.md
@@ -32,5 +32,6 @@ Fixed Webpack 4 issue, to explicity set mode to development.
 
 Added Sign Up Firebase Authentication.
 
-Added User Profile into Firebase Real-Time Database.
+Added Pills + Contact Doctors + Schedule Page.
 
+Added new pages above to app side bar navigation.